Migrate the subscription menu to Base UI (#893)
